inventory-service/pyproject.toml
[tool.poetry]
name = "inventory-service"
version = "0.1.0"
description = "Check stock"
authors = ["joelproxi <joeledmond95@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
django = "^5.1.3"
djangorestframework = "^3.15.2"
drf-yasg = "^1.21.8"
pika = "^1.3.2"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
